"Fix" how transaction names display in New Relic
New Relic's metric parser ruins how URIs are displayed in the browser,
as they use forward slashes to separate metric segments. With no way to
escape these slashes, it interprets the first part of the URL as a
"Controller name" (yes... as in a Rails controller), the second part of
the URL as the action name, and then discards the rest. That leaves us
with things like "GET#v1" if you've versioned your API via the URL path.
In order to "fix" this, we can use a Unicode character that looks like a
slash but actually isn't.

Signed-off-by: David Celis <me@davidcel.is>